requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
openpyxl>=3.1.0
xlrd>=2.0.0